--- blocksuite/affine/all/src/shared/viewport-renderer.ts
@@ -1,1 +1,0 @@
-export * from '@blocksuite/affine-shared/viewport-renderer';
