+++ blocksuite/affine/gfx/turbo-renderer/src/text-utils.ts
